chore(trading): remove waits and tests (#5637)
