+++ apps/trading/e2e/tests/positions/test_positions.py
@@ -15,7 +15,6 @@ def test_closed_market_position(vega: VegaServiceNull, page: Page):
         settlement_price=110,
         market_id=market_id,
     )
-    vega.forward("10s")
     vega.wait_fn(1)
     vega.wait_for_total_catchup()
     page.goto(f"/#/markets/{market_id}")
